+++ CHANGES.rst
@@ -14,6 +14,8 @@ Unreleased
   on poll submissions.
 * Add method :meth:`~.Reddit.delete` to :class:`.Reddit` class to support HTTP
   DELETE requests.
+* Added :class:`.CalendarConfiguration` to represent the configuration of a
+  :class:`.Calendar` widget.
 
 **Changed**